refactor(ui): 更新按键配置包路径
