+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/FtcLogFilesInspect.java
@@ -16,7 +16,7 @@ import org.betastudio.ftc.util.ButtonProcessor;
 import org.betastudio.ftc.util.message.TelemetryMsg;
 import org.betastudio.ftc.selection.SelectElement;
 import org.betastudio.ftc.selection.SelectPackage;
-import org.firstinspires.ftc.teamcode.ButtonConfig;
+import org.betastudio.ftc.util.ButtonConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
